Set default name for strategy, fix typo
While trying to use this gem with the setup_phase of omniauth for
dynamically assigning strategy options, I noticed that the name field
was defaulting to OpenIDConnect instead of openid_connect, which is
what the default route would be for omniauth, since the name isn't set
in the options until after it is needed in this case. Providing a
default name if one isn't set fixes that problem.

Regarding the typo. options.dicovery needed to be added, otherwise
public_key was being set to nil when using discovery with alternative
server signing algorithms.